Separated tests from production code
